+++ test/fixtures/critical.js
@@ -1,0 +1,2 @@
+/* eslint no-console: ["off"] */
+console.log('Critical JS loades successfully (must only be printed once!)');